a_projects/lsopc_backend/requirements.txt
uvicorn
lxml
playwright
fastapi
pydantic
pydantic-settings
requests
python-multipart
motor
pyjwt
bcrypt==4.0.1
passlib[bcrypt]
email-validator
